% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPP-package.R
\docType{package}
\name{MPP-package}
\alias{MPP-package}
\alias{MPP}
\title{MPP package}
\description{
Utility functions for marked point process models.
}
\author{
\strong{Maintainer}: Hengde Ouyang \email{houya@uic.edu}

Authors:
\itemize{
  \item Hengde Ouyang \email{houya@uic.edu}
}

}
